requirements.txt
requests==2.28.2
python-dotenv==0.21.1
pyyaml==6.0
requests-oauthlib==1.3.1 